CMake: Blanket PREFER_SYSTEM & USE_SYSTEM options
X-Moto allows certain libraries to be provided either by the system or
X-Moto's source tree itself, as vendored dependencies. (Not all vendored
dependencies may be provided by the system, such as Chipmunk.) These are
called vendorable libraries here.

With a USE_SYSTEM_* option ON, that vendorable library is system. With a
USE_SYSTEM_* option OFF, that vendorable library may be either system or
vendored.

With the USE_SYSTEM option OFF, all USE_SYSTEM_* options are OFF.
With the USE_SYSTEM option ON, all USE_SYSTEM_* options default ON.

With a PREFER_SYSTEM_* option ON, that vendorable library is searched
for and is system if found.

With the PREFER_SYSTEM option ON, all PREFER_SYSTEM_* options default
ON. With the PREFER_SYSTEM option OFF, all PREFER_SYSTEM_* options are
OFF.

With the USE_SYSTEM option OFF, the PREFER_SYSTEM option defaults ON.
With the USE_SYSETM option ON, the PREFER_SYSTEM option is OFF.

The upshot of all this is that USE_SYSTEM ON means libraries are
explicitly either system or vendored, with failed system searches
erroring instead of falling back to vendored, and both USE_SYSTEM &
PREFER_SYSTEM OFF means all libraries are vendored.

This should be good for maintainers?